remove useless packages
