+++ requirements_dev.txt
@@ -5,7 +5,6 @@ mock==2.0.0
 moto==1.3.5
 pytest-cov==2.5.1
 pytest==3.6.4
-Sphinx==1.7.6
 testing.postgresql==1.3.0
 tox==3.1.2
 bumpversion==0.5.3
